requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Background Tasks & Scheduling
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4

# Email & Notifications
fastapi-mail==1.4.1
jinja2==3.1.3

# Barcode & Product Data
python-barcode==0.15.1
Pillow==10.2.0
requests==2.31.0

# API & Data Validation
httpx==0.26.0
validators==0.22.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Monitoring & Logging
python-json-logger==2.0.7

# CORS
fastapi-cors==0.0.6